fix possible nullpointerexception

+++ catroid/src/org/catrobat/catroid/utils/UtilFile.java
@@ -87,11 +87,13 @@ public class UtilFile {
 	static public boolean clearDirectory(File path) {
 		if (path.exists()) {
 			File[] filesInDirectory = path.listFiles();
-			for (File file : filesInDirectory) {
-				if (file.isDirectory()) {
-					deleteDirectory(file);
-				} else {
-					file.delete();
+			if (filesInDirectory != null) {
+				for (File file : filesInDirectory) {
+					if (file.isDirectory()) {
+						deleteDirectory(file);
+					} else {
+						file.delete();
+					}
 				}
 			}
 		}
